Added buttonsWithReply | test
